# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Student {
  id: string;
  name: string;
  degree: string;
  year: string;
  careerGoal: string;
  location: string;
  workMode: string;
  stipendPreference: string;
  interests: string[];
  skills: string[];
}

export interface CreateStudentInput {
  id?: string;
  authUserId?: string;
  name: string;
  degree: string;
  year: string;
  careerGoal: string;
  location: string;
  workMode: string;
  stipendPreference: string;
  interests?: string[];
  skills?: string[];
}

export interface Internship {
  id: string;
  title: string;
  company: string;
  description: string;
  domain: string;
  location: string;
  workMode: string;
  duration: string;
  stipend: string;
  education?: string;
  requiredSkills: string[];
  preferredSkills: string[];
  experienceLevel: string;
}

export interface Recommendation {
  studentId: string;
  internshipId: string;
  score: number;
  reasons: string[];
  skillGap: string[][];
  internship: Internship;
}

export type InteractionAction = typeof InteractionAction[keyof typeof InteractionAction];


export const InteractionAction = {
  view: 'view',
  save: 'save',
  skip: 'skip',
  apply: 'apply',
  like: 'like',
} as const;

export interface Interaction {
  id: string;
  studentId: string;
  internshipId: string;
  action: InteractionAction;
  reason?: string;
  createdAt: string;
}

export type CreateInteractionInputAction = typeof CreateInteractionInputAction[keyof typeof CreateInteractionInputAction];


export const CreateInteractionInputAction = {
  view: 'view',
  save: 'save',
  skip: 'skip',
  apply: 'apply',
  like: 'like',
} as const;

export interface CreateInteractionInput {
  studentId?: string;
  internshipId: string;
  action: CreateInteractionInputAction;
  reason?: string;
}

export type GetRecommendationsParams = {
student_id?: string;
};

export type ListInteractionsParams = {
student_id?: string;
action?: string;
};