# lib/api-zod/src/generated/types/createStudentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudentInput {
  id?: string;
  authUserId?: string;
  name: string;
  degree: string;
  year: string;
  careerGoal: string;
  location: string;
  workMode: string;
  stipendPreference: string;
  interests?: string[];
  skills?: string[];
}